refactor: separate combat map and shop modules
