+++ tests/test_shop.py
@@ -1,4 +1,5 @@
 import os
+import os
 import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
@@ -6,6 +7,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")
 from dungeoncrawler.dungeon import DungeonBase
 from dungeoncrawler.entities import Player
 from dungeoncrawler.items import Item, Weapon
+from dungeoncrawler import shop as shop_module
 
 
 def test_shop_purchase(monkeypatch):
@@ -13,7 +15,7 @@ def test_shop_purchase(monkeypatch):
     dungeon.player = Player("Buyer")
     dungeon.player.gold = 20
     monkeypatch.setattr("builtins.input", lambda _: "1")
-    dungeon.shop()
+    shop_module.shop(dungeon)
     assert dungeon.player.gold == 10
     assert any(item.name == "Health Potion" for item in dungeon.player.inventory)
 
@@ -25,7 +27,7 @@ def test_sell_weapon(monkeypatch):
     dungeon.player.collect_item(weapon)
     inputs = iter(["1", "y"])
     monkeypatch.setattr("builtins.input", lambda _: next(inputs))
-    dungeon.sell_items()
+    shop_module.sell_items(dungeon)
     assert dungeon.player.gold == 20
     assert weapon not in dungeon.player.inventory
 
@@ -37,7 +39,7 @@ def test_sell_item(monkeypatch):
     dungeon.player.collect_item(potion)
     inputs = iter(["1", "y"])
     monkeypatch.setattr("builtins.input", lambda _: next(inputs))
-    dungeon.sell_items()
+    shop_module.sell_items(dungeon)
     assert dungeon.player.gold == 5
     assert potion not in dungeon.player.inventory
 
@@ -48,6 +50,6 @@ def test_sell_unsellable(monkeypatch):
     rare_weapon = Weapon("Elven Longbow", "Bow", 15, 25, 0)
     dungeon.player.collect_item(rare_weapon)
     monkeypatch.setattr("builtins.input", lambda _: "1")
-    dungeon.sell_items()
+    shop_module.sell_items(dungeon)
     assert dungeon.player.gold == 0
     assert rare_weapon in dungeon.player.inventory
